requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
grobid-client-python==0.0.4
h11==0.14.0
idna==3.4
isodate==0.6.1
lxml==4.9.2
ndg-httpsclient==0.5.1
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pip==22.0.4
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
rdflib==6.2.0
requests==2.28.2
selenium==4.7.2
setuptools==57.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tika==2.6.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wheel==0.38.4
wsproto==1.2.0
altgraph==0.17.3
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
isodate==0.6.1
lightrdf==0.3.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
rdflib==6.2.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
